DefectronQ/requirements.txt
Flask
flask-cors
torch
torchvision
Pillow
pennylane
numpy
